git-ffintegrateto: CHG: Minor: Push superproject branch before deleting submodule branch(es)
The order doesn't really matter; swapping has the advantage that the check for automatic branch deletion by GitHub (which takes a few seconds if it happened and a good 20 seconds if not - for each submodule) happens while a GitHub action is already working on the pushed superproject branch; the time is now counting towards the GitHub action completion and does not delay it any longer.

FIX: Minor: When isPerformTheMerge is cleared, we should exit with $status; previous push failures will clear this flag when they fail.