gitconfig-br-variants: ENH: br..l[o]by now only offer authors that have committed in the range
Instead of selecting all authors that have committed in previous commits.
The range has to be passed both to the eventual l[o]g and (new) also to others-command, which now understands --since (in the timespan commands) and --range RANGE (all others).
This will also fix mismatches if a custom --branch (or --revision) is given.