ENH: Add git-dpls, dpls[1-7xqzg], and br..dpsl[g] variants
These are similar to sls in that they first ask the user for a (here: single, because the iteration of multiple diffs from multiple commits could be confusing) commit, but then show each changed file separately (git-dpl) instead of showing each (of multiple commits) log + diff separately.
Note: We could take a --count=<number> and pass that to git-dp, but the -<number>|-n <number> variants would clash with git-showselected, and I don't think diffing over multiple commits is a valid use case here.