git-unmerge: ENH: Also handle fast-forward integrated branches
The merge happens on the branch (i.e. the target (master) branch is merged into it instead of the other way around), so the merge commit is on the branch itself.
That means that if we're already on that branch, the merge commit needs to be dropped, and that's it (we'll ignore whether something happened on the target branch already, and assume the user did a "git owipe && git colb" or so already).
Else, check out and reset the branch.