ENH: Add reffintegrate command to reattempt a fast-forward integration after the push of the target branch was rejected
As I mostly rely on my ffintegrateto command for merges, picking up the pieces is a hassle when the final push of the fast-forwarded master branch (in the superproject) fails because someone else managed to merge another branch between the start of my reintegration and that push.
The script distinguishes between history-preserving merge and single-commit rebase, and handles everything but the fetching.