hub-delete-merged-branches: Force-delete remote+local branches, too
When the superproject integration fails due to conflicts or incompatibilities introduced by the other branch, the merged branch needs to be amended for the fix.
This somehow caused Git to complain about the local branch not being fully merged, and the hub-supersubreintegratetom to abort. The user confirmed the deletion, so go ahead with force.