BUG: "git subdo --dirty --untracked" does not work
Because predicates add more conditions, so this will only catch submodules with both untracked and modified files. Making the combination (default: &&) of predicates configurable would be one solution, but cumbersome and complex if it is / can be changed during parsing.
What we actually need is a combined predicate; unfortunately, a good name for "untracked || dirty" is hard to find: "unstaged" definitely excludes files in the index (which "dirty" includes) and (at least in my use) excludes files not under source control. Same for "uncommitted".
I've decided to use "stageable" for untracked or modifications not yet in the index, which makes this not a direct combination (as it excludes the stage).
Add git-stageable variant of git-untracked (i.e. requiring --print to show the files, unlike git-dirty), --stageable and stageable-sh to git-do-extensions.