git-remote-mutate: ENH: Check whether the remote branch already is up-to-date
And then just print a note instead of querying the user whether they want to push.
This is skipped as soon as any other arguments are present, so a -f|--force will override this check. (This may need to be necessary when the remote has been updated through another repo and the current repo hasn't been fetched yet, but that's a rare case and in general it's recommended to fetch first.)
A push not covered by git-remote-mutate or when --yes is given will still push, even if there are no changes, but that's okay, since the annoying thing is the query, not the short delay from the network operation itself.
But the main motivation for adding this is in my submodules workflow, where a confirmation on a reintegrated submodule that got fast-forwarded (and therefore no push is necessary as there's no merge commit) always caused confusion (and frequent use of the "Show context" help). Now the command directly says: Hey mate, there's nothing to do here.