ENH: Add cor[s], cors[s1-7xqzg] and br...cops variants that allow selection of commit from <range> first, then FILE(s) (cp. 9c71027)
cop[s] is for [FILE(s) from] the commit prior to the passed / selected one, but targeting the commit directly also is a valid use case. This so far was also covered by cos and wipe (with additional side effects like adding to the index), but it's better to have a dedicated command for that, and parallelism between cop and cor.

(This relates to commit 9c710273ce87ad7e5ab9bbc88b39e2db7c08cd2c)