ENH: Add trb... variant to the commit before the last rebase
Similar to tm..., but targeting the commit (from the reflog) from before a rebase was started.
Some commands (esp. diff) are a bit different, and most of the commands likely aren't of much use.
However, trbd is great to ensure that a rebase of fixups actually ends up with the same end result (when manual merge conflict resolution occurred) - for this, I previously had to somehow track the original commit (look it up in the reflog, or push the prior branch state and use outd, etc.)
trbbaselg can now directly show the original commit, and trbwipe could be an "undo rebase".