BUG: Cannot pass a SIMPLECOMMAND containing a -- inside
The first -- so far aborts the command parsing and the remainder is taken as FILE|DIR args; that's bad, as SIMPLECOMMAND may itself want to use --. One option would be a --exec SIMPLECOMMAND [...] ; alternative, but that requires client changes.
Mandate a -- before FILE|DIR if SIMPLECOMMAND itself contains a --, and only interpret the args following the last -- as FILE|DIR.
When no -- is encountered, parse last arguments for existing files.