FIX: Outgoing submodule reintegration commits are always found
Even when the reintegration would be a fast-forward, because when the check is run, the submodule(s) have already been reintegrated to master (and will be pushed once the superproject has been merged and checked).
Therefore, "git outlg" (on master) will always contain commits, at least the ones from the branch now waiting to be pushed (in case of a fast-forward the original ones, but if a rebase has happened different ones), plus optionally the merge commit.
In order to distinguish between fast-forward vs. rebase or merge, we need to obtain the diff between the previous (reintegrated) branch and its (still existing, the delete of submodule branches happens only later) upstream.
"git blbase" will give us the previous branch name; as we don't know whether we're invoked through the $submoduleExecutor (which supports --command CMDLINE) or plain git-wrapper (which does not), we'll do a recursive self-invocation with the special --print-outgoing instead of embedding the command substitution (which has to happen inside the submodule (well, if we assume that the same branch name is used everywhere we could also just query the superproject's current branch)).