git-supersubreintegratetom: ENH: Offer integration in GitHub after local merge and checks
Note: Logically we could have just checked for possible hub-integrate first and then decided based on the existence of outgoing reintegration commits whether full delegation to hub-integrate is possible or local work needs to be done.
However, this would do the remote check for a PR twice (first for checking and then for doing). Also, the usability is slightly better this way because the fact that there are outgoing reintegration commits is printed prior to the offer. Downside is some duplicated code.