git-up: UI: Cosmetics: Pipe the after-update check messages through hushup as well
I usually use git-up in combination with git-hushup, and the processing of the update operation's output then interferes with any error message from the after-update check; i.e. the check gets printed _before_ the update messages themselves. This is not only confusing (temporal inversion), but also leads to a missing separation between the after-update message and the "From github.com:..." of the update.
For git-submodule-uninitialized, I can use the gitUpdateOperation prefix; for the after-update checks, I need to use an explicit unconditional redirection to be able to use outputAndPrintf.
CHG: DWIM: Don't stop after the first check fails, run them all.
Inclusion of the check messages in stdout has the side effect that git-hushup attributes them to the last submodule fetch (if any happened), which is bad. To avoid that, inject a custom "Fetching done." message that's only included if a check fails. git-hushup uses this message to end the submodule fetching part (and drops the message, which hides the fact that it's printed for each failing check).