FIX: @{-1} does not reference the integrated branch and it would be wrong, anyway (after 90409ba)
After submodule reintegration, the branch would be updated in case of a rebase, but a merge commit would not be found there, only on the (still unpushed) master. Therefore, the approach of comparing the previous (reintegrated) branch and its (still existing, the delete of submodule branches happens only later) upstream is faulty.
And in any case it turned out that "git blbase" somehow does not give us the previous branch name.
But we know from the superproject's remote branch which submodule commit is current in the remote, and have to compare that with the unpushed current submodule state. In the case of a rebase, commits have been altered, in case of a merge there's a new merge commit and commits from the other parent, only in a fast-forward there are no changes and the superproject reintegration could proceed in and by GitHub.

(This corrects commit 90409bac0794fb5d91c836b5637e210c46114b77)