git-remote-mutate: FIX: Handling of (third-arg) <refspec> wrong (after 3174a5a)
For "git push origin foo", we need to obtain the tracked branch for "foo", and then compare "$remote/foo" with "foo". Instead we used the current branch's name and compared with HEAD, but <refspec> actually specifies (in this simple form) what source branch is pushed, not necessarily the current branch.
Fortunately, this makes the logic (that was wrong, too (what was I thinking?) and was previously corrected) even more simple.

(This corrects commit 3174a5a8945ddc14a514a6a292ba93b9ab74381e)