ENH: Add commit{identicalto,like}super
Directly referencing the last / passed revision in the superproject.
Note: There are no commitrelate variants the dependency order typically is from the superproject to the submodule, not the other way around (so we'd rather need a commitrelatetosubmodule command, but there's the complication that it's a one-to-many relationship).
Note: There are no correct variants because a submodule typically cannot correct something done in the superproject.