CHG: [uo]uptodate now reference the tracked remote branch and ignore the state of the local branch
So ouptodate answers the question whether I need to fetch origin changes from remote, but not whether the current branch is even with the tracked remote branch. Previously, both questions were combined (but git-uptodate directly checks the remote even if there's a discrepancy already between the current branch and the tracked remote branch, for which no remote repo access would be necessary).