hub-workflow-status: BUG: Regression: Leaving behind parts of the message after waiting concludes (after 75b2836)
Noticed in a git-supersubffintegratetom run:
> Waiting for the GitHub action to validate the build on the branch before pushing.
> 10:02: The fast-forward-integrated branch build got completed.
> ess  3 minutes ago     https://github.com/XxxxxxxxxXxxxx/XxXxxxx/actions/runs/70
> Final up-to-date check before pushing main branches quasi-transactionally:

I removed --quiet but forgot to drop the --silent-command flag for retry as well. The flag can now only be used (and do that, as it's nicer visually without the clearing) if --quiet is passed.

(This corrects commit 75b28368a7ade760d7a2b380fe2e6004f53dd6c4)