hub-delete-merged-branches: BUG: Regression: Revision lookup fails because nothing is passed (after a59e140)
Need to pass the revision in $branch!
Robustness: Exclude the current branch, as that would be unexpected as a target for deletion (and likely cause problems). Unfortunately, after a fast-forward integration the previous branch (@{-1}) and the current (master) branch are on identical commits, so it's likely that the current branch is returned!

(This corrects commit a59e14003fce33f63d53a560ce3a62bcaf935cd2)