git-reintegrate: BUG: hub-supersubffintegratetom of fast-forwarded submodules fails
Because the submodules are detached when "Fast-forwarding submodule into main".
The --no-submodule-update flag that hub-supersubffintegratetom passes via git-ffintegrateto to git-reintegrate must also apply to the fast-forwarding. Any mergeArgs passed by the user (likely a merge strategy if it's ever done) do no harm when fast-forwarding, so we don't have to split off another array but can simply pass all mergeArgs to any merge.