hub-workflow-status: ENH: Add --wait-for-success variant of --wait-for-completion
That additionally checks that the conclusion is "success". So a completed workflow run that failed (e.g. due to a build error) will not cause the loop to fail. The conclusion ("failure") will be shown in the wait loop output (and the user'll likely receive a GitHub notification separately), so that the user can go and restart the job.
Adapt: Reintegrations need to wait for a successful workflow run, as only that ensures that the reintegrated master branch can be pushed. Add --concluded success.
CHG: Switch -W to stand for --wait-for-success and introduce -w for --wait-for-completion. (--wait is so short and meant for scripting that it doesn't require a short option). With this, the documentation in hubcheats-submodules-manually.md doesn't need updating as it uses the short -W.