CHG: Renaming: clone[d]br to stack[d]br, pc.../...topc to ps.../tops
I found the "stacked branches" description at https://andrewlock.net/working-with-stacked-branches-in-git-is-easier-with-update-refs/ and like it!
Cloning is the low-level technical term (and I wasn't fully happy with that as there's overlap with the much more prominent "git clone" of repos).
It's better to describe this via the high-level end result: a stack of numbered branches that is supposed to be applied one after the other.