Robustness: Don't amend to the previous superproject commit when all submodules have been integrated but not committed
Other problems had happened before, and when I tried to get the submodule references updated to a multi-commit superproject branch, it mistakenly interpreted the situation as an 10th of 9 submodules and amended, instead of created the desired separate housekeeping commit:

,----[ example ]----
| $ hub supersubffintegratetom
| Checking for unintegrated submodules... No such submodules found.
| All 9 submodules have been reintegrated already; now integrating the superproject.
| Apparently the submodule fast-forward-integration was interrupted and 9 submodule references have not been committed yet.
| Update the changed submodule references in the superproject now? ([Y]es / [N]o)? y
|
| XxXxxxx:
|  M XxXxxxxxxxx
|  M XxXxxxx
|  M XxXxxxxx
|  M XxXxxx
|  M XxXxXx
|  M XxXxxx
|  M XxXxxxxxx
|  M XxXxxxxxxxx
|  M XxXxxxxx
| Amending the 10. reintegrated submodule to the previous superproject commit.
| [tests/refactoring/Translations a5dd6f2] Tests: Refactoring: Extract TranslationCorrectnessFixture into XxXxxx and apply to all plugins
`----
This then led to follow-up problems (like the push failing because --force was missing but the amend had changed an existing commit). I should have switched to manual commands from that point in time...

The problem is caused by the interruption detection invoking completeSubmoduleReintegrationInSuperproject() but that function not handling this special situation.
Add a check for pendingSubmoduleCnt > #affectedSubmodules (which effectively is the same as #unintegratedSubmodules == 0, and cannot happen in the normal invocation after a submodule reintegration).