CHG: DWIM: Check out an existing (local or remote) reference by default
Use case: I started working on a feature, then someone else pulled my branch and added commits. Now if I only pulled their changes in the superproject, but leave the submodule's branches, a checkout of the branches would rewind the submodule changes.
Instead, what I want is a checkout of submodule branches only if the superproject directly references them. For the branches, it shouldn't matter whether my local or the remote branch references it; in case of the latter, my local branch should be updated.
The superproject reference is the arbiter; if that points to the tip of a branch, checkout that branch, but if it references any other commit, this means that the active development has ceased and we should respect that reference.