git-subcobr: FIX: Check for start point also within the submodule
The (symbolic) REV may only exist in the superproject, but not in (some) submodules. Better check where REV will be referenced; i.e. in the submodule. This also makes the check in isSkipped() consistent with the other checks, which also just use the flags set by getSubmoduleStatus() instead of doing queries on their own.