Robustness: Abort if a submodule reintegration fails
The failure has important information on continuation that should not get lost in the sea of output.
Also, a failure is a stressful situation, so it's best to interrupt the flow rather than show even more potential failures in following submodules. The submodule reintegration can be easily continued (we can reintegration submodule by submodule even), so this is not disadvantage at all.