Robustness: Abort if a reintegration in and by GitHub fails
Here, the user has explicitly chosen the GitHub reintegration, so we shouldn't fall back to reintegrating locally (or even a mixture of remote and local reintegrations).
Therefore, abort on the first failure and exit.
It's different (and more difficult to differentiate) for the offering; there, the user has to decline following questions if one submodule fails.