Robustness: Detect fast-forward superproject fails (e.g. due to check failure) and offer resumption of submodules, wait, and push
In a fast-forward integration, the submodules and then the superproject are reverse-integrated, then the checks run. If those fail, the user needs to fix the problem (e.g. in a submodule), then amend to the (submodule) merge commit, force-push the branch, amend to the superproject commit, and then continue the failed superproject integration: $ hub ffintegrateto --continue
That runs the checks again, and if they now pass, we're at a state where the submodule branches are reverse-integrated, and the superproject is reverse-integrated and fast-forwarded to master already. We still need to fast-forward the submodules, wait for the GitHub action, then push everthing. That's a lot to do manually, and we're also losing the quasi-transactionality when pushes are done manually, or we'd have to craft the subdo command that does it all in one go. It would be much better if git-supersubffintegratetom could simply be reinvoked and resume where it left off.
Due to the submodules still on the source branch, the script still recognizes the affected submodules (with a minor tweak for --same-branch). We can detect this situation then through the superproject being on master, and then offer to complete the remainder, which already is neatly encapsulated in fastforwardCommit().