FIX: Misleading continuation message when checks fail and a merge has happened
Suggesting a restart is counterproductive when a merge has happened; the script does not detect that and instead will attempt to create another merge on top.
A failure after a merge should be tackled by amending the fix to the merge, even making additional commits is not recommended.
I don't fully recollect the rationale for the restart, and don't understand this any longer:
> If the current branch has been modified, the reintegration need to be restarted to be able to fast-forward. We cannot simply continue after the original integration commit has been left behind.
For rebases, the restart doesn't do much less harm, but I think if we just did amends or additional commits, a complete restart isn't necessary, as long as the remote side didn't become updated.
Only when a remote branch required updates (fixes on the target branch, forgotten commits on the tracked source branch in origin), the rebase needs to be restarted.

Add a conditional to give specialized recommendations for rebases vs. merges.
Offer the continuation command first, as this is the more likely choice. The restart command is only offered as an alternative.