git-ffintegrateto: BUG: Continuation if the superproject is rebased fails because no force-push
When the submodule(s) merge commit(s) are amended to the single superproject commit, the branch needs to be force-pushed. The script automatically handles that by checking $isRebase in pushBranch(). However, when the command is continued (after a build failure), the reintegration logic that determines whether a rebase is done is skipped, and $isRebase is false, so no --force is supplied to "git push".
Fortunately, there's --rebase to explicitly choose rebasing, so printContinueCommand() just needs to print that if $isRebase is set.