git-do-extensions: FIX: Only parse the first -- into dashdashArgs, and no args after the second --
git-wcdo is documented as [WCDO-ARGS ...] [--] WORKING-COPY [...] [-- SIMPLECOMMAND ...], but many of my wrappers (e.g. Unixhome-ingo) pass WORKING-COPY -- [WCDO-ARGS ...] [--] [SIMPLECOMMAND] (and to make matters worse through git-wcs-in-dir-do-command). I now need to reuse git-do-extensions in git-wcs-in-dir-do (to handle the extension's added flags like --dirty), and for that need proper (or at least parseable) separation between wcdoArgs, dashdashArgs, and args.
When everything that's not a wcdo-arg is parsed into dashdashArgs (all assuming GIT_DOEXTENSIONS_DASHDASH_ARGS_BEFORE_SIMPLECOMMAND=t, of course), that's not helpful, as we'd have to do all the parsing ourselves once more.
I hope this change will serve git-wcs-in-dir-do while maintaining compatibility with git-wcdo.