git-wcs-in-dir-do: FIX: Extension options like --dirty are not understood
For extension commands like dirty-sh, this is no problem; these have been forwarded to wcdo, which then uses git-do-extensions to convert them into stuff it understands.
But for options like --dirty, we'd have to duplicate all of them here in addition to the existing duplicate wcdo options. Better get rid of that duplication and delegate the parsing to git-do-extensions.