hub-integrate: Robustness: Handle unset default repo for gh
This isn't nice:
> ERROR: This branch '...' is tracking origin, but GitHub is set up to interact with branch '' instead.

First problem is that gh only performs the repo query when output is to a terminal (and doesn't even print a warning then), but hub-integrate uses command substitution to obtain the remote's name. Give a (similarly worded) error ourselves.
Second problem is that hub-integrate only aborts because the conditional is false as $remote is empty. Add a dedicated check; we don't need an error message as hub-remote now provides this.