Robustness: Complain when passed SUBMODULE [...] haven't been found
The existing check is more a consistency assertion between git-sublist (the source of existingSubmodules) and the .gitmodules contents. Exit 3 if there's a mismatch.
But if the user passed an invalid SUBMODULE, this isn't caught. Instead, remove each printed submodule from selectedSubmodules, and complain in the end if there are left-overs.