git-ffintegrateto: BUG: Need to consider --mark-reverse-merge-done after --continue, too
When a super-sub combination is fast-forward-integrated via git-supersubffintegratetom and the submodule reintegration fails (in this case a merge conflict due to a binary file), the latter script needs the marker for the reverse merge.
Without that marker, the restarted git-supersubffintegratetom does not recognize the submodules reintegration and attempts to do that again:
> Integrating the single submodule.
The marking is an integral part of the fast-forward-integration; add it to continueArgs so that it doesn't get lost.