WIP: Suppress the merge suggestion from ffintegrateto when invoked in a submodule from supersubffintegratetom
When merging happens automatically, this hint is just a nuisance; each submodule is automatically reintegrated by supersubffintegratetom. But in case of a merge conflict or other failure, this misleads the user into executing it, which is not how the fast-forward-process happens.
In a fast-forward-integration, we want the submodules to first stay on the (fast-forward-integrated) branch, and only will do the fast-forwarding later as part of the quasi-transactional push.
Add a --no-merge-suggestion flag (also passed to a command continuation) that suppresses the final hint.