git-supersub{ff,re}integratetom: Documentation: Add comprehensive help on continuation after a problem
Based on the last failures, these steps hopefully correct any problems encountered during (ff-)reintegration. It should be enough to just follow the hints printed by the individual commands, but as this is a complex process and usually under stressful conditions, it's helpful to have a written process description to cross-reference and use to orient oneself.