hub-subpull-request: ENH: Include the superproject commits in the commit message template as well
We already include the superproject's commit messages, as a potential source for the submodule's PR title and/or description.
For an exclusive submodule PR, it may make sense to include references to the (minor, otherwise we should open a dedicated superproject PR) commits as well. This is analog to the superproject referencing submodule commits if there's only a superproject PR.
The new git-superchanges opposite of git-subchanges gives us that information; splice it into the superMessage. (Note: In contrast to git-subchanges this doesn't handle PR titles at all, because then we'd have a super+submodule constellation, and hub-supersubpull-request would be used, anyway.)