ENH: Add correcttostash and stashrelate variants of correct / commitrelate
Stashed changes may also want to refer to existing commits, either to correct then (when it's inconvenient to directly commit a correction now), or just to relate to another commit.
With git-stash-push, we've allowed the drop-in replacement of git-commit, and just need a tiny extension point in git-correct to make it happen.
git-correcttostash and git-stashrelate are very thin wrappers that use this new extension point.
gitconfig has all the alias variants based on correct and commitrelate, with the additional twist that the *i variants only stash the index and the *-i variants only stash the working copy, which I hope is DWIM.