git-commit: ENH: Allow commit message templates in .git/commit-template-COMMITID
I already have a similar mechanism implemented in git-reintegrate and git-ffintegrateto; this extends this idea into my commit wrapper itself (but I'll have to keep the existing implementations as they are for git-merge's commiting, not git commit directly).
This allows a command like "git stash pop" to seed a commit message that is then used by a following commit.