git-stash-{apply,pop}: ENH: Save the stash's message as a commit template, so that it can be reused for the commit
With git-commit now picking up a saved commit message template, we can take a (meaningful) stash message and seed the commit message that happens after a "git stash pop|apply" with it.
This completes the use case of using git correcttostash / stashrelate: When stashing, context is saved in the stash message, and that is then reused in the commit message once the stash is applied and committed.