ENH: Add rbsedreword[1-7xqz] variants
This is a bit tricky, as we cannot define a rbsedreword alias; the <range> would have to be placed between the included sedreword and its user-provided arguments somehow.
Likewise, the rbsedrewordselected needs a new --no-file-args to avoid that the (mandatory for git-rebase) -- SIMPLECOMMAND is treated as FILE(s).
We also cannot simply delegate to rbsedrewordselected for the rbsedreword[1-7xqz] variants; the range needs to be placed inside.
Because of these complications, I refrain from defining br... variants here. The coarser numeric ranges should be fine for selection of commit ranges, as this is about the commit messages, and their subjects are visible in the selection dialog, anyway.
I also could define rbpipereword[1-7xqz] variants, but that's also too exotic to consider right now.