hub-supersubpull-request: Insert the commit subjects and subchanges between PR title and existing body
The commit subjects are meant as templates for title replacement, and the subchanges will be at the top of the pull request body.
Note: There's a performance penalty in splitting the PR retrieval into two calls, but keeping one call and splitting it myself, and then passing that into tempedit which takes source commands isn't trivial, neither.