git-stash-pop: ENH: DWIM: Detect dirty index and dual stash and automatically supply --index
The default "git stash pop" requires the user to supply --index to apply a single stash directly to the index, or a dual stash to both index and working copy. How do I know?
I'd rather have the command do the right thing(tm) and apply a dual stash to both (if possible), or warn about the conflict with the index.
A single stash application to the index will fail (but strangely drop an existing staged new file) or create a conflict; I'd rather have a warning here.