ENH: Push a merge to a special integration branch
If we push the fast-forward merge to the source branch, a pull request for that one will also show all commits that have happened on the master branch, and the merge commit itself, which is confusing. I'd rather have the pull request maintain the state from prior to the merge.
To achieve this, we can push to a (remote-only, no need for a local branch; source: https://stackoverflow.com/a/52338714/813602) new integration branch (named after the source branch but with -integration appended).
For the checks performed by a GitHub workflow, what matters is the commit ID, not on which branch this was submitted.