git-commitcheck: ENH: Show one combined log of the highlighted problems
It's unlikely that different problems occur, but if it does, the sequential logging takes up too much space.
Prepend the commit ID to the commit subject, adapt the grep to anchor to the ^G separator, and then extract the commit ID.
Use the git-lghighlightedcommits for highlighting. With this, we'll also correctly handle a --grep log option (which are ORed whereas we'd need AND).