git-{[uo]push,rpublish}: ENH: Check for fixup and temporary commits when pushing
This is just a note (and we rely on a git-remote-mutate confirmation or the user being fast enough to press Ctrl+C instead of confirming). Fixup commits may only be squashed later after PR review is done, and sometimes it's necessary to push test commits to the CI/CD pipeline.