git-{sub,wc}do-core: ENH: Allow presetting the color mode via GIT_{SUB,WC}DOCORE_COLOR
I have checkUnixhome that needs to force (git-wcdo; i.e. for the header and executed Git command) color mode, but it invokes {Unixhome,vimrc}-{originals,ingo}, and those invoke git-wcs-in-dir-do-command [...] -- "$@", so any passed --color=always is not at the right position (it would need to be before the --).
I don't want to extend all those various wrappers with full command-line argument parsing (which would be the right thing to do). Dropping the -- from the wrapper would mean that during interactive use, the -- had to be passed (e.g. Unixhome-ingo -- ufetch), which is bad.
Let's add an environment variable through which the checkUnixhome can preset the color mode.