git-rblastfixup: ENH: Also handle out-of-order fixups
The previous implementation assumed that the first fixup commit also references the earliest commit target, but that isn't necessarily so; the last fixup can also target the oldest commit.
We need to resolve the target commits individually for each fixup (as it's the most recent with that commit subject, but there could be previous commits with an identical subject).
If there are multiple target commits, git-logrevs puts them into the correct order and we can pick the earliest from them.