git-addlike: Filter out non-existing files
If we simply pass a non-existing file (i.e. one that got removed in the previous commit), the whole "git add" aborts with
> fatal: pathspec '...' did not match any files

I could use "git showfiles --not-deleted", but we might want to re-add a previously deleted file. It's better to actually filter the file list according to what's now there.