Fall back to replacement on full branch if branch prefix / suffix doesn't change it
For branch creation / renaming, we may want to change the prefix / suffix itself, but as that's excluded, it won't work.
Clearing the environment variable(s) is cumbersome and unintuitive.
Attempting a fallback is a 50%-solution; it won't help with ^= and += (as those will always change the branch name), but allows us to rename inside the {pre,suf}fix via OLD=NEW.