git-{subdo,subrev,supersubdo,wcdo}*: FIX: Only extract the SIMPLECOMMAND's color arg if no --no-git-color and no --color yet
In my dashboard/etc/6099work-wips.sh, I enable coloring for git-wcdo but need to disable coloring for the git-wips SIMPLECOMMAND. git-wcdo must not parse the later --no-color and take that for itself, as I've explicitly given --color=always to it.
With an added --no-git-color, I can avoid that git-wcdo passes its own color as well (though that would result in --color=always --submodule-color=always --no-color, so the latter wins, which is what I want).