hub-su{per,b}pull-request: ENH: Allow just printing the PR message template with --print-message
This is useful when commits have been added or rebased after the PR was opened. (Parts of) the message can then be copied and pasted (in GitHub's web editor or gh pr edit --body).
Note: This isn't added to hub-supersubpull-request due to its more complex logic, and because the other two variants can be used to obtain targeted messages for either superproject or submodules.