completions: Usability: Filter out count suffixes and variants first to reduce the amount of completions
I now have so many aliases; many commands have default counts ([1-7xqzg]) and some come in variants for *[st,i,-i]; and the completion candidate list often is overwhelming.
Suppress counts as long as we're not yet at the aliases' base itself.
Suppress variants as long as we're not yet at that variant's base itself.