ENH: Offer remote to push to if no tracked remote branch yet
We may not want to publish the branch previously, as that will already trigger a CI/CD build, and for trivial changes, the build of the fast-forward merge is enough. (And in fact, it may be problematic to immediately trigger two separate builds after another.)
By querying the user, we allow for aborting the thing, and also handle multiple remotes.