git-correct: ENH: Optionally reference a tag in the title if there's a single matching tag
A (version) tag is more expressive than a (short) commit ID.
Introduce a (positive-only) tag glob configuration similar to what's used in git-tagtype; this serves both as an opt-in and a necessary filter so that temporary or non-evocative tags (like archive/*) aren't referenced.
If multiple tags are attached, it might be better to ask the user for a choice, but right now I'm simply bailing and falling back to the ID.
For the "This corrects ...", we always want the full commit ID. This change is only about the reference in the title.