hub-subpull-request: FIX: The title candidates must not be Markdown-escaped
Split this into the commit message titles from the superproject (unescaped), followed by the Markdown-escaped raw subjects and bodies. Submodule-related commits can be skipped here, so use "git superchanges" here as well. (The following "git superchanges --as markdown" serves a different purpose; it's not a template, but provides a summary of the superproject changes.)