git-supersubreintegratetom: Documentation: Document how to clear the reverse-merge mark
I had to leave behind a started (but failing) supersubffintegration; on the next day, another submodule integration had added a commit on master.
My single submodule commit had already been rebased on master the day before, and so the script deemed this as integrated and skipped it. But I needed another rebase to pick up that new commit.