Documentation: Cross-reference commitkeywordtally and whatdid
I had the feeling that I'd implemented something like that already...
The use case is a bit different, though, so it's okay to keep both:
There are already br...whatdid variants, and these are quick to use interactively as the logging is built-in.
commitkeywordtally is meant for automated reporting by git-devstat, and the separation of the logging means that it can report across multiple repos.